Move fetchOrgAndApps from dev/fetch to PartnersClient
It's Partners-specific logic, so that's where it belongs.
